requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.37.33
botocore==1.37.33
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2024.10
cryptography==44.0.2
Django==4.2.20
django-allauth==65.7.0
django-appconf==1.1.0
django-compressor==4.5.1
django-cors-headers==4.7.0
django-countries==7.6.1
django-crispy-forms==2.3
django-debug-toolbar==5.1.0
django-filter==25.1
django-redis==5.4.0
django-storages==1.14.6
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
flower==2.0.1
gunicorn==23.0.0
humanize==4.12.2
idna==3.10
jmespath==1.0.1
kombu==5.5.2
packaging==24.2
paypalrestsdk==1.13.3
pillow==11.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
rcssmin==1.1.2
redis==5.2.1
requests==2.32.3
rjsmin==1.2.2
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
stripe==12.0.0
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
